backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
openai==1.23.2
python-dotenv==1.0.1
python-multipart==0.0.9
